requirements.txt
# Django==1.9.4
# djangorestframework==3.3.3
djangorestframework==3.14.0
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
django-cors-headers>=4.0,<5.0   # 如果前后端分离需要 CORS
pandas>=2.0,<3.0                # CSV/Excel 处理
openpyxl>=3.1,<4.0              # Excel 读写
requests